move ht_op_to_channel_width to common code
This will be used by both ampe and meshd at some point, so move
it out of meshd.
